wip: setup logging before uacfg is loaded
