+++ lib/patch_status_json.py
@@ -19,7 +19,7 @@ import json
 import logging
 
 from uaclient import system, util
-from uaclient.cli import setup_logging
+from uaclient.log import setup_logging
 
 
 def patch_status_json_schema_0_1(status_file: str):
