+++ lib/timer.py
@@ -6,7 +6,6 @@ import logging
 from datetime import datetime, timedelta, timezone
 from typing import Callable, Optional
 
-from uaclient.cli import setup_logging
 from uaclient.config import UAConfig
 from uaclient.exceptions import InvalidFileFormatError
 from uaclient.files.state_files import (
@@ -17,6 +16,7 @@ from uaclient.files.state_files import (
 from uaclient.jobs.metering import metering_enabled_resources
 from uaclient.jobs.update_contract_info import update_contract_info
 from uaclient.jobs.update_messaging import update_motd_messages
+from uaclient.log import setup_logging
 
 LOG = logging.getLogger(__name__)
 UPDATE_MESSAGING_INTERVAL = 21600  # 6 hours
