+++ uaclient/tests/test_cli.py
@@ -2,11 +2,8 @@ import contextlib
 import io
 import json
 import logging
-import os
 import re
 import socket
-import stat
-import sys
 import textwrap
 
 import mock
@@ -21,7 +18,6 @@ from uaclient.cli import (
     assert_root,
     get_parser,
     main,
-    setup_logging,
 )
 from uaclient.entitlements import get_valid_entitlement_names
 from uaclient.exceptions import (
@@ -684,7 +680,7 @@ class TestMain:
         assert "Traceback (most recent call last):" in error_log
 
     @pytest.mark.parametrize("caplog_text", [logging.DEBUG], indirect=True)
-    @mock.patch("uaclient.cli.setup_logging")
+    @mock.patch("uaclient.cli.pro_log.EarlyLoggingSetup")
     @mock.patch("uaclient.cli.get_parser")
     def test_command_line_is_logged(
         self, _m_get_parser, _m_setup_logging, logging_sandbox, caplog_text
@@ -737,182 +733,6 @@ class TestMain:
             )
             == str(err)
         )
-
-
-class TestSetupLogging:
-    @pytest.mark.parametrize("level", (logging.INFO, logging.ERROR))
-    @mock.patch("uaclient.cli.util.we_are_currently_root", return_value=False)
-    def test_console_log_configured_if_not_present(
-        self, m_we_are_currently_root, level, capsys, logging_sandbox
-    ):
-        setup_logging(level, logging.INFO)
-        logging.log(level, "after setup")
-        logging.log(level - 1, "not present")
-
-        _, err = capsys.readouterr()
-        assert "after setup" in err
-        assert "not present" not in err
-
-    @mock.patch("uaclient.cli.util.we_are_currently_root", return_value=False)
-    def test_console_log_configured_if_already_present(
-        self, m_we_are_currently_root, capsys, logging_sandbox
-    ):
-        logging.getLogger().addHandler(logging.StreamHandler(sys.stderr))
-
-        logging.error("before setup")
-        setup_logging(logging.INFO, logging.INFO)
-        logging.error("after setup")
-
-        # 'before setup' will be in stderr, so check that setup_logging
-        # configures the format
-        _, err = capsys.readouterr()
-        assert "ERROR: before setup" not in err
-        assert "ERROR: after setup" in err
-
-    @mock.patch("uaclient.cli.util.we_are_currently_root", return_value=False)
-    def test_file_log_not_configured_if_not_root(
-        self, m_we_are_currently_root, tmpdir, logging_sandbox
-    ):
-        log_file = tmpdir.join("log_file")
-
-        setup_logging(logging.INFO, logging.INFO, log_file=log_file.strpath)
-        logging.info("after setup")
-
-        assert not log_file.exists()
-
-    @pytest.mark.parametrize("log_filename", (None, "file.log"))
-    @mock.patch("uaclient.cli.config")
-    def test_file_log_configured_if_root(
-        self,
-        m_config,
-        log_filename,
-        logging_sandbox,
-        tmpdir,
-    ):
-        if log_filename is None:
-            log_filename = "default.log"
-            log_file = tmpdir.join(log_filename)
-            m_config.CONFIG_DEFAULTS = {"log_file": log_file.strpath}
-        else:
-            log_file = tmpdir.join(log_filename)
-
-        setup_logging(logging.INFO, logging.INFO, log_file=log_file.strpath)
-        logging.info("after setup")
-
-        assert "after setup" in log_file.read()
-
-    def test_file_log_configured_if_already_present(
-        self,
-        logging_sandbox,
-        tmpdir,
-    ):
-        some_file = tmpdir.join("default.log")
-        logging.getLogger().addHandler(logging.FileHandler(some_file.strpath))
-
-        log_file = tmpdir.join("file.log")
-
-        logging.error("before setup")
-        setup_logging(logging.INFO, logging.INFO, log_file=log_file.strpath)
-        logging.error("after setup")
-
-        content = log_file.read()
-        assert re.match(r'\[.*"ERROR", "before setup"', content) is None
-        assert re.match(r'\[.*"ERROR",.*"after setup"', content)
-
-    @mock.patch("uaclient.cli.config.UAConfig")
-    def test_custom_logger_configuration(
-        self,
-        m_config,
-        logging_sandbox,
-        tmpdir,
-        FakeConfig,
-    ):
-        log_file = tmpdir.join("file.log")
-        cfg = FakeConfig({"log_file": log_file.strpath})
-        m_config.return_value = cfg
-
-        custom_logger = logging.getLogger("for-my-special-module")
-        root_logger = logging.getLogger()
-        n_root_handlers = len(root_logger.handlers)
-
-        setup_logging(logging.INFO, logging.INFO, logger=custom_logger)
-
-        assert len(custom_logger.handlers) == 2
-        assert len(root_logger.handlers) == n_root_handlers
-
-    @mock.patch("uaclient.cli.config.UAConfig")
-    def test_no_duplicate_ua_handlers(
-        self,
-        m_config,
-        logging_sandbox,
-        tmpdir,
-        FakeConfig,
-    ):
-        log_file = tmpdir.join("file.log")
-        cfg = FakeConfig({"log_file": log_file.strpath})
-        m_config.return_value = cfg
-        root_logger = logging.getLogger()
-
-        setup_logging(logging.INFO, logging.DEBUG)
-        stream_handlers = [
-            h
-            for h in root_logger.handlers
-            if h.level == logging.INFO and isinstance(h, logging.StreamHandler)
-        ]
-        file_handlers = [
-            h
-            for h in root_logger.handlers
-            if h.level == logging.DEBUG
-            and isinstance(h, logging.FileHandler)
-            and h.stream.name == log_file
-        ]
-        assert len(root_logger.handlers) == 2
-        assert len(stream_handlers) == 1
-        assert len(file_handlers) == 1
-
-        setup_logging(logging.INFO, logging.DEBUG)
-        stream_handlers = [
-            h
-            for h in root_logger.handlers
-            if h.level == logging.INFO and isinstance(h, logging.StreamHandler)
-        ]
-        file_handlers = [
-            h
-            for h in root_logger.handlers
-            if h.level == logging.DEBUG
-            and isinstance(h, logging.FileHandler)
-            and h.stream.name == log_file
-        ]
-        assert len(root_logger.handlers) == 2
-        assert len(stream_handlers) == 1
-        assert len(file_handlers) == 1
-
-    @pytest.mark.parametrize("pre_existing", (True, False))
-    @mock.patch("uaclient.cli.config")
-    def test_file_log_is_world_readable(
-        self,
-        m_config,
-        logging_sandbox,
-        tmpdir,
-        pre_existing,
-    ):
-        log_file = tmpdir.join("root-only.log")
-        log_path = log_file.strpath
-        expected_mode = 0o644
-        if pre_existing:
-            expected_mode = 0o640
-            log_file.write("existing content\n")
-            os.chmod(log_path, expected_mode)
-            assert 0o644 != stat.S_IMODE(os.lstat(log_path).st_mode)
-
-        setup_logging(logging.INFO, logging.INFO, log_file=log_path)
-        logging.info("after setup")
-
-        assert expected_mode == stat.S_IMODE(os.lstat(log_path).st_mode)
-        log_content = log_file.read()
-        assert "after setup" in log_content
-        if pre_existing:
-            assert "existing content" in log_content
 
 
 class TestGetValidEntitlementNames:
